[docker-compose] now generate config file at start
this is the first step to integrate the number of nodes per roles when
using the command start.
For docker-compose, we generate the docker-compose.json as previously
for the composition.
when starting, we will extract the information about the number of nodes
for each role and rewrite the docker-compose file to take them into
account