Adding a command line option to nxc connect for a custom SSH key.
By default, it runs an SSH command without specifying the key to be used. Therefore, if one has several identity keys, SSH may fail.
I am afraid it is a solution to one's specific problem which many would not face, and a poor approach at it. It consists indeed of adding some text to the SSH command line. Feels poorly integrated.