[Example] Changed nxcShellFull to nxcShell as the naming convention changed.
Note to devs: this change was not made in the other examples,
leading to the ubiquitous use of nxcShell (and absence of
nxcShellLite) in examples. This is a deliberate choice to avoid
any potential bug introduced by removing tmux support or any
other feature. Feel free to downgrade nxcShell to nxcShellLite
in some examples if possible.